New Forgot Password Function in Login
For simplicity and cost (email servers for 2fa are expensive and this is just a small app), this uses google's app password, compatible with google accounts.

A user can request a code through the login screen, with either a username or email. To reduce spam, limited only to those with the team access code. A secure token is sent to their inbox, with a link, which also expires. The user enters the new password, and logs in through a new page.